refactor: Reporter/HTML/Reporter to Reporter/Reporter.
